refactor: rename scope-reference-this util to scope-references-this
Caught by filenames/match-exported lint rule.
