+++ lib/rules/no-lazy-arrow-functions.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const isIdentifier = require('../utils/is-identifier');
-const scopeReferencesThis = require('../utils/scope-reference-this');
+const scopeReferencesThis = require('../utils/scope-references-this');
 const isTestFile = require('../utils/is-test-file');
 
 function isLazy(node) {
