+++ lib/utils/scope-references-this.js
